Fixed nav bar on edit account

+++ src/main/java/edu/wpi/teamb/controllers/components/InfoCardController.java
@@ -8,7 +8,9 @@ import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
 import javafx.scene.control.Alert;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.Label;
+import javafx.scene.control.Tooltip;
 import javafx.scene.image.ImageView;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.VBox;
@@ -17,6 +19,7 @@ import javafx.stage.Stage;
 
 import java.io.IOException;
 import java.util.Objects;
+import java.util.Optional;
 
 public class InfoCardController {
   @FXML private AnchorPane requestInfoAnchorPane;
@@ -30,6 +33,7 @@ public class InfoCardController {
   @FXML private Label timeSubmittedLabel;
   @FXML private Label locationNameLabel;
   @FXML private Label employeeAssignedLabel;
+  @FXML private Label requestTypeLabel;
   @FXML private Label statusLabel; //deprecated implementation comment: make a button for this *Note: Please Disregard*
   @FXML private VBox buttonContainerVBox;
   @FXML private VBox subComponentContainer;
@@ -51,6 +55,7 @@ public class InfoCardController {
   }
 
     private void initBtns() {
+      completeButton.setTooltip(new Tooltip("Click to mark request as Completed"));
         completeButton.setOnMouseClicked(e -> {
             if (!fullRequest.getRequestStatus().equals("Completed")) {
                 //update fullRequest status
@@ -76,15 +81,22 @@ public class InfoCardController {
                 alert.showAndWait();
             }
         });
-
+        deleteButton.setTooltip(new Tooltip("Click to delete request"));
     deleteButton.setOnMouseClicked(
         event -> {
-          //remove the request from the list of requests
-            requestInfoAnchorPane.setVisible(false);
-            ((VBox) requestInfoAnchorPane.getParent()).getChildren().remove(requestInfoAnchorPane);
-            //delete the request from the database
-            EInfoCard.deleteRequest(fullRequest);
+            Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+            alert.setTitle("Delete Edge");
+            alert.setContentText("Are you sure you want to delete this request?");
+            Optional<ButtonType> action = alert.showAndWait();
+            if (action.get() == ButtonType.OK) {
+                //remove the request from the list of requests
+                requestInfoAnchorPane.setVisible(false);
+                ((VBox) requestInfoAnchorPane.getParent()).getChildren().remove(requestInfoAnchorPane);
+                //delete the request from the database
+                EInfoCard.deleteRequest(fullRequest);
+            }
         });
+    editButton.setTooltip(new Tooltip("Click to edit request"));
     editButton.setOnMouseClicked(event -> {
         FXMLLoader loader = null;
         Scene editPageScene = null;
@@ -169,8 +181,19 @@ public class InfoCardController {
     if (fullRequest.getDateSubmitted() != null) {
         //break up the date and time into two different variables
         String[] dateAndTime = fullRequest.getDateSubmitted().toString().split(" ");
-        setDateSubmittedLabel(dateAndTime[0]);
-        setTimeSubmittedLabel(dateAndTime[1]);
+        String[] date = dateAndTime[0].split("-");
+        String dateFormatted = date[1] + "/" + date[2] + "/" + date[0];
+        setDateSubmittedLabel(dateFormatted);
+        String[] time = dateAndTime[1].split(":");
+        String timeFormatted;
+        if (Integer.parseInt(time[0]) > 12) {
+            timeFormatted = Integer.toString(Integer.parseInt(time[0]) - 12) + ":" + time[1] + " PM";
+        } else if (Integer.parseInt(time[0]) == 0) {
+            timeFormatted = "12" + ":" + time[1] + " AM";
+        } else {
+            timeFormatted = time[0] + ":" + time[1] + " AM";
+        }
+        setTimeSubmittedLabel(timeFormatted);
     } else {
         setDateSubmittedLabel("Unavailable");
         setTimeSubmittedLabel("Unavailable");
@@ -179,6 +202,7 @@ public class InfoCardController {
     setEmployeeAssignedLabel(fullRequest.getEmployee());
     setStatusLabel(fullRequest.getRequestStatus());
     requestTypeIconImageView.setImage(fullRequest.setRequestTypeIconImageView());
+    requestTypeLabel.setText(fullRequest.getRequestType());
     fullRequest.setRequestType();
   }
